move db stuff
